// Laudari.WebApp/laudari-webapp/src/app/core/models/ultrassom/aorta/aorta.enum.ts
export enum AortaClassificacao {
    Normal = 0,
    Ateromatosa = 1,
    AteromatorsaEctasiada = 2,
    Aneurisma = 3,
    Inacessivel = 4
}

export enum AortaAneurisma {
    Fusiforme =  0,
    Secular = 1
}

export enum AortaLocalAneurisma {
    Infrarenal =  0,
    TercoMedio = 1,
    TercoProximal = 2,
    TercoDistal = 3,
    JuntoBifurcacao = 4
}